Fix int overflow in Smooth/Smooth3 catch-up loop counter
Change loop variable from int to long in both Smooth and Smooth3
getCurrentValue() catch-up loops. The delta is a long, so an int
counter would overflow and loop infinitely if delta exceeded
Integer.MAX_VALUE.

+++ courant-engine/src/main/java/systems/courant/sd/model/Smooth.java
@@ -159,7 +159,7 @@ public class Smooth implements Formula, Resettable {
                 }
                 st = 1.0;
             }
-            for (int i = 0; i < delta; i++) {
+            for (long i = 0; i < delta; i++) {
                 smoothed += (inputVal - smoothed) / st;
             }
             lastStep = step;
